Added simplecov to git-checkout
