+++ .gitignore
@@ -1,1 +1,4 @@
 *.swp
+errors.err
+.cadre
+coverage
