+++ git-checkout/lib/git-checkout.rb
@@ -1,0 +1,3 @@
+module GitCheckout
+	autoload :App, 'git-checkout/app'
+end
